requirements.txt
selenium
webdriver-manager
pyperclip
fastapi
uvicorn
psutil
filelock
pyautogui
Pillow